Fixed IINC; the value was read as unsigned, but the jvm specs require it to be signed. This caused issues with negative increment values. I have no doubt this same issue exists in other instructions...

+++ JavaAsm/Instructions/InstructionListConverter.cs
@@ -268,7 +268,7 @@ namespace JavaAsm.Instructions {
                         else {
                             instructions.Add(currentPosition, new IncrementInstruction {
                                 VariableIndex = codeStream.ReadByteFully(),
-                                Value = codeStream.ReadByteFully()
+                                Value = (sbyte) codeStream.ReadByteFully()
                             });
                         }
